Clean up filename handling - move it to bash classify.sh runs end-to-end

+++ Assignment 3 - Graph Classification/scripts/test.sh
@@ -1,17 +1,22 @@
 #!/bin/bash
-run(){
-	./libraries/svm-train -t 0 -c 500 train.txt model_linear.txt
-	./libraries/svm-predict test.txt model_linear.txt predicted.txt
-	python scripts/fscore.py "$1" predicted.txt
-	rm model_linear.txt predicted.txt
-}
+if [ "$#" == 4 ]; then
+    TRAIN_FILE_NAME="${1%.*}"
+    TRAIN_FILE_EXT="${1##*.}"
+    TRAIN_FILE_CONV="$TRAIN_FILE_NAME"_converted."$TRAIN_FILE_EXT"
+    TRAIN_FILE_LAB="$TRAIN_FILE_NAME"_labels."$TRAIN_FILE_EXT"
+    TEST_FILE_NAME="${4%.*}"
+    TEST_FILE_EXT="${4##*.}"
+    TEST_FILE_CONV="$TEST_FILE_NAME"_converted."$TEST_FILE_EXT"
+    TEST_FILE_LAB="$TEST_FILE_NAME"_labels."$TEST_FILE_EXT"
 
-if [ "$#" == 3 ]; then
-    python graphToVec.py "$1" "$2" --support "$3"
-    run "$2"
-elif [ "$#" == 2 ]; then
-    python graphToVec.py "$1" "$2"
-    run "$2"
+    python scripts/format.py "$1" "$TRAIN_FILE_CONV" --labels "$TRAIN_FILE_LAB" --active "$2" --inactive "$3"
+    python scripts/format.py "$4" "$TEST_FILE_CONV"
+    python graphToVec.py "$TRAIN_FILE_CONV" "$TRAIN_FILE_LAB" "$TEST_FILE_CONV" "$TEST_FILE_LAB"
+
+    ./libraries/svm-train -t 0 -c 500 train.txt model_linear.txt
+    ./libraries/svm-predict test.txt model_linear.txt predicted.txt
+    python scripts/extra/fscore.py "$TEST_FILE_LAB" predicted.txt
+    rm model_linear.txt predicted.txt
 else
     echo "Illegal number of parameters ($#)"
 fi
